rename logging.py to log.py to avoid import error in py2
