+++ bentoml/__init__.py
@@ -30,7 +30,7 @@ from bentoml.service import (
 from bentoml.server import metrics
 from bentoml.archive import save, load
 
-from bentoml.utils.logging import configure_logging
+from bentoml.utils.log import configure_logging
 
 configure_logging()
