+++ bentoml/artifact/pytorch_model_artifact.py
@@ -19,6 +19,7 @@ from __future__ import print_function
 import os
 
 from bentoml.artifact import ArtifactSpec, ArtifactInstance
+from bentoml.utils import cloudpickle
 
 
 class PytorchModelArtifact(ArtifactSpec):
@@ -42,12 +43,13 @@ class PytorchModelArtifact(ArtifactSpec):
         except ImportError:
             raise ImportError("torch package is required to use PytorchModelArtifact")
 
-        from bentoml.utils import cloudpickle
         model = cloudpickle.load(open(self._file_path(path), 'rb'))
 
         if not isinstance(model, torch.nn.Module):
-            raise TypeError("Expecting PytorchModelArtifact loaded object type to be "
-                            "'torch.nn.Module'")
+            raise TypeError(
+                "Expecting PytorchModelArtifact loaded object type to be "
+                "'torch.nn.Module'"
+            )
 
         return self.pack(model)
 
@@ -75,5 +77,4 @@ class _PytorchModelArtifactInstance(ArtifactInstance):
         except ImportError:
             raise ImportError("torch package is required to use PytorchModelArtifact")
 
-        from bentoml.utils import cloudpickle
         return cloudpickle.dump(self._model, open(self.spec._file_path(dst), "wb"))
